Move list module tests to separate file
